Change 'i.quotes' width
It's been changed to '60%' of the parent container(quotes-container),
so I'll have to erase the centralization attributes from the
'.content.mid' selector and place it in the '.quotes-container'. It'll
be because, since I attributed a relative size, the position is now
related to the imediate parent container(I suppose). This is for my
future self.